Language refactor 1: Show first-time usage info and otherwise keep UI simple.
For now we'll assume that UI language == content language, and we'll make
that clear to the user. We might add additional language UI later, but
right now it's not clear that the bang for the buck is there.

This commit adds a generic dismissible notice type, including temporary
dismissal (via JS only) and permanent dismissal (via JS or form request),
with a nice example of how to handle the same request using a web route
or an API route, without code duplication. We now have req.isAPI to
check for the request type.

We're also now sanely returning errors processing API requests ..
through the API.