Language refactor 2: Reviews to multilingual strings, resolved in templates.
It made no sense to have the model overwrite its own values, or even to
store this in the model at all. Instead we use a template helper for
this, which also adds a nice language indicator next to each
string that is not in the current language (if desired).

With reviews using mlstrings, we now have architectural support
for translations of reviews. Yay!