Languages refactor 3, messages refactor
Switch to POST request for changing languages. This protects those
requests with a CSRF token, so we can't trick users into changing
their language preference (we might add a ?uselang= query parameter
to allow changing it for just one page, with appropriate
notification).

This also refactors the way system messages and errors are handled,
keeping them in their own flash key, and showing them on every page.
pageErrors/pageMessages on the other hand are only flashed when
needed.

We're adopting the bold box style for all messages except errors
that are likely to be stacked, such as form errors.