Support for creating/editing/deleting teams.
This also introduces the new FormHandler class, which should make
CUD forms easier to build. We'll need to refactor the review form
to make use of it.

We've also enhanced the form parser, so it can now handle escaping
and markdown conversions.

Finally, the revisioning helper now supports initial revision
creation.