Add support for ::: spoiler, ::: nsfw and ::: warning (foo) blocks
This is a custom extension to markdown using the recommended
container plugin. We're using the HTML5 <details> tag for
expand/collapse blocks, but also progressively enabling/enhancing it
with JS/CSS where available.

We're treating the UI of these content blocks as content, i.e.
it's added to the parsed and stored HTML. Since we retain the source
we can change that if needed, but since we allow custom warning
texts, these are really more content-like than UI-like.

A few linter issues fixed along the way.