IRC bot: automatically regain canonical nickname
Make the bot check every thirty seconds that its nickname matches its
configuration. If it does not, attempt to regain the canonical nickname
via "regain".

Note that this is Freenode-specific. It could be made more generic by
replacing the `bot.say(...)` line with `bot.send('NICK', config.botName)`,
but using NickServ is more powerful, because it will force anyone else
who may be holding the nick to release it.