Pretty URLs, round 2: short URLs for review subjects ('things')
As with teams, we create short human-readable strings from
labels, but in this case we also auto-disambiguate in case there
is more than one thing with the same name. For now we do so by
appending a counter, but in future we may use metadata such as
author, year, type of work, etc. to generate qualifiers.