Add search
This adds a JavaScript-based autocompletion search feature, as well
as a full text/title search that works on all browsers. It
depends on ElasticSearch (tested with 5.2.0) being up and running,
and search-related features will fail if that isn't the case.

Once this is stable, we should add ES to the Vagrant role as well.